Ignore redundant task_struct argument when p_ed_process is provided
There are several functions which take both a task_struct pointer and a
p_ed_process pointer. This presents a hazard if the provided task_struct
doesn't match the provided p_ed_process, with ensuing mayhem as a result.

To prevent this from ever happening, refactor all such functions to either
remove the redundant task_struct argument or make it optional. When a
function accepts both a task_struct and p_ed_process, the task_struct
pointer stored inside p_ed_process is always preferred; if the p_ed_process
argument is NULL, then the task_struct argument is used.

Signed-off-by: Sultan Alsawaf <sultan@ciq.com>